Backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-extensions==3.2.3
django-ninja==1.1.0
pydantic==2.7.0
pydantic_core==2.18.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
